updated ksp function and set nix-kube-cluster to only be on non-Windows OS-es

+++ dotfiles/config/powershell/not-windows.ps1
@@ -38,14 +38,21 @@ function mosh() {
 
 # Append the nix-kube-cluster config to the current kubeconfig
 function nix-kube-cluster {
-    bw get notes nix-kube-cluster/config > ~/.kube/nix-cluster-config
-
-    $env:KUBECONFIG="$HOME/.kube/config:$HOME/.kube/nix-cluster-config"
-    kubectl config view --flatten > ~/.kube/config-merged
-    rm ~/.kube/config
-    mv ~/.kube/config-merged ~/.kube/config
-    $env:KUBECONFIG="~/.kube/config"
-    rm ~/.kube/nix-cluster-config
+    # Only move forward if nix-kube context does not exist
+    if (-not (kubectl config get-contexts | grep -iq nix-kube)) {
+        New-Item -ItemType Directory -Path ~/.kube -ErrorAction SilentlyContinue
+        bw get notes nix-kube-cluster/config > ~/.kube/nix-cluster-config
+        if (-not (test-path ~/.kube/config)) {
+            $env:KUBECONFIG="$HOME/.kube/nix-cluster-config"
+        } else {
+            $env:KUBECONFIG="$HOME/.kube/config:$HOME/.kube/nix-cluster-config"
+        }
+        kubectl config view --flatten > ~/.kube/config-merged
+        rm ~/.kube/config
+        mv ~/.kube/config-merged ~/.kube/config
+        $env:KUBECONFIG="$HOME/.kube/config"
+        rm ~/.kube/nix-cluster-config
+    }
 }
 
 function start-ssh-agent () {
